// dashboard/src/api/models/taxIdTypeEnum.ts
export type TaxIdTypeEnum = (typeof TaxIdTypeEnum)[keyof typeof TaxIdTypeEnum];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const TaxIdTypeEnum = {
  unknown: "unknown",
  ad_nrt: "ad_nrt",
  ae_trn: "ae_trn",
  al_tin: "al_tin",
  am_tin: "am_tin",
  ao_tin: "ao_tin",
  ar_cuit: "ar_cuit",
  au_abn: "au_abn",
  au_arn: "au_arn",
  aw_tin: "aw_tin",
  az_tin: "az_tin",
  ba_tin: "ba_tin",
  bb_tin: "bb_tin",
  bd_bin: "bd_bin",
  bf_ifu: "bf_ifu",
  bg_uic: "bg_uic",
  bh_vat: "bh_vat",
  bj_ifu: "bj_ifu",
  bo_tin: "bo_tin",
  br_cnpj: "br_cnpj",
  br_cpf: "br_cpf",
  bs_tin: "bs_tin",
  by_tin: "by_tin",
  ca_bn: "ca_bn",
  ca_gst_hst: "ca_gst_hst",
  ca_pst_bc: "ca_pst_bc",
  ca_pst_mb: "ca_pst_mb",
  ca_pst_sk: "ca_pst_sk",
  ca_qst: "ca_qst",
  cd_nif: "cd_nif",
  ch_uid: "ch_uid",
  ch_vat: "ch_vat",
  cl_tin: "cl_tin",
  cm_niu: "cm_niu",
  cn_tin: "cn_tin",
  co_nit: "co_nit",
  cr_tin: "cr_tin",
  cv_nif: "cv_nif",
  de_stn: "de_stn",
  do_rcn: "do_rcn",
  ec_ruc: "ec_ruc",
  eg_tin: "eg_tin",
  es_cif: "es_cif",
  et_tin: "et_tin",
  eu_oss_vat: "eu_oss_vat",
  eu_vat: "eu_vat",
  gb_vat: "gb_vat",
  ge_vat: "ge_vat",
  gn_nif: "gn_nif",
  hk_br: "hk_br",
  hr_oib: "hr_oib",
  hu_tin: "hu_tin",
  id_npwp: "id_npwp",
  il_vat: "il_vat",
  in_gst: "in_gst",
  is_vat: "is_vat",
  jp_cn: "jp_cn",
  jp_rn: "jp_rn",
  jp_trn: "jp_trn",
  ke_pin: "ke_pin",
  kg_tin: "kg_tin",
  kh_tin: "kh_tin",
  kr_brn: "kr_brn",
  kz_bin: "kz_bin",
  la_tin: "la_tin",
  li_uid: "li_uid",
  li_vat: "li_vat",
  ma_vat: "ma_vat",
  md_vat: "md_vat",
  me_pib: "me_pib",
  mk_vat: "mk_vat",
  mr_nif: "mr_nif",
  mx_rfc: "mx_rfc",
  my_frp: "my_frp",
  my_itn: "my_itn",
  my_sst: "my_sst",
  ng_tin: "ng_tin",
  no_vat: "no_vat",
  no_voec: "no_voec",
  np_pan: "np_pan",
  nz_gst: "nz_gst",
  om_vat: "om_vat",
  pe_ruc: "pe_ruc",
  ph_tin: "ph_tin",
  ro_tin: "ro_tin",
  rs_pib: "rs_pib",
  ru_inn: "ru_inn",
  ru_kpp: "ru_kpp",
  sa_vat: "sa_vat",
  sg_gst: "sg_gst",
  sg_uen: "sg_uen",
  si_tin: "si_tin",
  sn_ninea: "sn_ninea",
  sr_fin: "sr_fin",
  sv_nit: "sv_nit",
  th_vat: "th_vat",
  tj_tin: "tj_tin",
  tr_tin: "tr_tin",
  tw_vat: "tw_vat",
  tz_vat: "tz_vat",
  ua_vat: "ua_vat",
  ug_tin: "ug_tin",
  us_ein: "us_ein",
  uy_ruc: "uy_ruc",
  uz_tin: "uz_tin",
  uz_vat: "uz_vat",
  ve_rif: "ve_rif",
  vn_tin: "vn_tin",
  za_vat: "za_vat",
  zm_tin: "zm_tin",
  zw_tin: "zw_tin",
} as const;